Add linux support
Add paths and scripting for to support operation on linux.

I wasn't able to get this to build, but it seems it's some issue with
how I'm invoking pyinstaller.  As a follow-on to this, I'll add a commit
that moves a couple things around to allow the build to succeed which
can be considered separately from this one.